Moved logging middleware file
