+++ internal/server/middleware/logging.go
@@ -1,4 +1,4 @@
-package server
+package middleware
 
 import (
 	"fmt"
